avl/requirements.txt
asyncpg>=0.29.0
python-dotenv>=1.0.0

# Web tracker
flask>=3.0.0
psycopg2-binary>=2.9.9
